[Step4] refactor : 인터페이스 패키지 이동
