+++ src/main/kotlin/blackjack/domain/player/Dealer.kt
@@ -2,7 +2,7 @@ package blackjack.domain.player
 
 import blackjack.domain.card.CardDeck
 import blackjack.domain.game.TakeMoreDealer
-import blackjack.domain.game.TakeMoreDealerStrategy
+import blackjack.domain.game.strategy.TakeMoreDealerStrategy
 
 class Dealer(cardDeck: CardDeck) : Player(name = DEALER_NAME, cardDeck) {
